requirements.txt
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain-community
tenacity==8.3.0
fastapi[all]